Rejigging
Main update has been to change the purpose of MongoDB. For this - I want it to be purely a location to save the data we have used. This meant changing where the data is uploaded to MongoDB, and this has been put in the app.py code. This has been done to align with the use of Docker